// client-monorepo/libs/shared/payment/transactions/src/lib/data-access/models/transaction-type.enum.ts
export enum TransactionType {
  PAYMENT_REQUEST = 0,
  PURCHASE = 1,
  UTILITY_BILL = 2,
  TOP_UP = 3,
  CARD = 4,
  ACH = 5,
  PAYMENT_OFFLINE = 6,
  INTERNET_PACKAGE = 7,
  IN_APP_PURCHASE = 10,
  REWARD = 11,
  REFUND = 12,
  WALLET = 13,
  CASH_IN = 14,
  TOLL = 15,
  TOLL_PAYOFF = 16,
  CREDIT_PURCHASE = 17,
  PAYOUT_CARD = 18,
  REFUND_MANUAL = 19,
  CONGESTION_PRICING = 20,
  REFUND_DIGIPAY = 21,
  REFUND_POS = 22,
  CASH_OUT = 23,
  THIRD_PARTY_TICKET = 24,
  REVERSE_MANUAL = 26,
  PAYMENT_LINK = 27,
  INSTALLMENT = 28,
  CREDIT_INQUIRY = 29,
  DONATION = 30,
  TRAFFIC_FINE = 31,
  PARTNER_AGENT_PAYOUT = 32,
  CREDIT_SPLIT_PURCHASE = 33,
  CREDIT_PREPAYMENT = 34,
  SUBSCRIPTION_INVOICE = 35,
  MARKETPLACE = 36,
  STOCK = 37,
  TAXI_PAYMENT = 38,
  VOUCHER = 39,
  DIGITAL_SIGNATURE = 40,
  MERCHANT_CREDIT_SETTLEMENT_FEE = 41,
  SPLIT_PURCHASE = 42,
  CREDIT_RESIDUE = 43,
  CREDIT_FILING = 44,
  TRAFFIC_FINE_INQUIRY = 45,
  CREDIT_ALLOCATION_PREPAYMENT = 46,
  REVOLVING = 47,
  BUSINESS_CASH_OUT = 48,
  BUSINESS_CASH_IN = 49,
  BUSINESS_WALLET_TRANSFER = 50,
  CREDIT_EARLY_SETTLEMENT = 51,
  PROMISSORY_NOTE_PAYMENT = 52,
  SHAHKAR_INQUIRY = 53,
  REGISTRATION_INFO_INQUIRY = 54,
  ADDRESS_INQUIRY = 55,
  CHEQUE_INQUIRY = 56,
  CORRECTIVE_PAYOUT = 57,
  CHEQUE_COLOR_INQUIRY = 58,
  PICHAK_INQUIRY = 59,
  PROMISSORY_NOTE = 60,
  PROMISSORY_NOTE_INQUIRY = 61,
  VOUCHER_COLLECT = 62,
  MERCHANT_CREDIT_REGISTRATION = 63,
  APP_SUBSCRIPTION = 64,
  ASSET_COLLATERAL_PAYMENT = 65,
  MERCHANT_SETTLEMENT = 66,
  MERCHANT_FEE = 67,
  MCI_SUBSCRIBER_INQUIRY = 68,
  CREDITWORTHINESS_INQUIRY = 69,
  IBAN_COMPLIANCE_INQUIRY = 70,
  APP_SUBSCRIPTION_ASSIGNED = 71,
  WEALTH_CASH_IN = 72,
  WEALTH_CASH_OUT = 73,
}